Change return type from `CGFloat` to `NSUInteger`

+++ Zxcvbn/DBUtilities.h
@@ -11,7 +11,7 @@
 
 @interface DBUtilities : NSObject
 
-+ (CGFloat)calcBruteForceCardinalityForPassword:(NSString *)password;
++ (NSUInteger)calcBruteForceCardinalityForPassword:(NSString *)password;
 
 + (NSString *)displayTimeForSeconds:(CGFloat)seconds;
